+++ packages/analyzer/src/features/framework-plugins/lit/static-properties.js
@@ -11,7 +11,7 @@ import { isAlsoAttribute, hasStaticKeyword, getPropertiesObject, getAttributeNam
 export function staticPropertiesPlugin() {
   return {
     name: 'CORE - LIT-STATIC-PROPERTIES',
-    analyzePhase({ts, node, moduleDoc}){
+    analyzePhase({ts, node, moduleDoc, context}){
       switch (node.kind) {
         case ts.SyntaxKind.ClassDeclaration:    
           const className = node?.name?.getText();
@@ -41,6 +41,9 @@ export function staticPropertiesPlugin() {
                   const attributeName = getAttributeName(property);
                   if(attributeName) {
                     attribute.name = attributeName;
+                    classMember.attribute = attributeName;
+                  } else {
+                    classMember.attribute = classMember.name;
                   }
 
                   if(reflects(property)) {
@@ -65,9 +68,7 @@ export function staticPropertiesPlugin() {
           });
 
           /** Get default values */
-          currClass?.members?.forEach(member => {
-            getDefaultValuesFromConstructorVisitor(node, member);
-          });
+          getDefaultValuesFromConstructorVisitor(node, currClass, context);
           break;
         }
       }